Moved window related structures to window manager
