+++ include/PipelineManager.hpp
@@ -8,7 +8,7 @@
 class PipelineManager {
 public:
   PipelineManager(std::unique_ptr<class VulkanContext> &vulkanContext,
-                  std::unique_ptr<class Raytracer> &raytracer);
+                  std::unique_ptr<class Raytracer> &raytracer, std::unique_ptr<class WindowManager>& window);
   ~PipelineManager();
 
   void createGraphicsPipeline(VkDevice device, VkRenderPass renderPass,
@@ -19,7 +19,7 @@ public:
   const VkRenderPass &getRenderPass() const;
   const VkFramebuffer getFrameBuffer(int i) const { return framebuffers[i]; }
   
-  void recreateFramebuffers(std::unique_ptr<VulkanContext>& vulkanContext);
+  void recreateFramebuffers(std::unique_ptr<VulkanContext>& vulkanContext, std::unique_ptr<WindowManager>& window);
 
   static VkShaderModule createShaderModule(VkDevice device,
                                            const std::vector<char> &code);
